delete randomsearch test
I was having an absurd amount of problems getting the scikeras wrapper and randomsearchcv to work, and not enough time to fix them.

I considered them not worth the hastle to include, since they are not even that much more performant than the other neural networks to make it worth it to keep and fix, so I have decided to remove them.